fix: corrige modal inversao cardapio
